test: Station 관련 요청 메서드 분리
